requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blis==1.3.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.13
confection==0.1.5
csvw==3.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
decorator==5.2.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
durationpy==0.10
edge-tts==7.2.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
executing==2.2.1
fastapi==0.119.1
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
genai==2.1.0
generativeai==0.0.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyboard==0.13.5
kokoro==0.9.4
kubernetes==33.1.0
langchain-chroma==0.2.6
langchain-core==0.3.76
langchain-huggingface==0.3.1
langcodes==3.5.0
langsmith==0.4.31
language-tags==1.2.0
language_data==1.3.0
Levenshtein==0.27.1
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
misaki==0.9.4
mmh3==5.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
networkx==3.5
num2words==0.5.14
numpy==2.3.3
nvidia-ml-py==13.580.82
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.0
openai==0.27.10
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
parso==0.8.5
phonemizer-fork==3.3.2
pillow==11.3.0
playsound==1.2.2
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
py3nvml==0.2.7
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybase64==1.4.2
pycaw==20240210
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pynput==1.8.1
pynvml==13.0.1
pyparsing==3.2.5
pyperclip==1.10.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-mpd2==3.1.1
pytweening==1.2.0
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.14.1
rdflib==7.2.1
redis==7.0.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
segments==2.3.0
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speedtest-cli==2.1.3
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
TgCrypto==1.2.5
thefuzz==0.22.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.3.3
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vosk==0.3.45
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0
WMI==1.5.1
wrapt==1.17.3
xmltodict==1.0.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
